requirements.txt
amqp==2.6.0
asgiref==3.2.10
billiard==3.6.3.0
celery==4.4.6
Django==3.0.7
django-appconf==1.0.4
django-celery-email==3.0.0
django-cors-headers==3.4.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
future==0.18.2
gunicorn==20.0.4
importlib-metadata==1.6.1
kombu==4.6.11
psycopg2==2.8.5
PyJWT==1.7.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
vine==1.3.0
zipp==3.1.0
